fix after review

+++ server/controllers/bank.controller.js
@@ -20,15 +20,28 @@ module.exports.getBanks = (req, res) => {
   let { page } = req.params;
   let { limitItems } = req.params;
 
+  limitItems = parseInt(limitItems);
   page = parseInt(page);
+  const skip = (page - 1) * limitItems;
 
-  Bank.find({ idOwner: userId }).exec((err, bank) => {
-    if (bank) {
-      result = bank.slice((page - 1) * limitItems, page * limitItems);
+  Bank.aggregate([
+    {
+      $match: {
+        idOwner: userId,
+      },
+    },
+    {
+      $skip: skip,
+    },
+    {
+      $limit: limitItems,
+    },
+  ]).exec((err, banks) => {
+    if (banks.length > 0) {
       res.status(200).send({
         success: true,
-        message: "[SUCCESS] User has banks",
-        data: result,
+        message: "[SUCCESS] User  has banks",
+        data: banks,
       });
     } else {
       res.status(202).send({
@@ -41,62 +54,46 @@ module.exports.getBanks = (req, res) => {
 };
 module.exports.getOneBank = (req, res) => {
   const { id } = req.params;
+  const { userId } = req;
 
-  Bank.findOne({ _id: id }).exec(async (err, bank) => {
-    let bankRes = {};
-    let numberOfEasyQuestions = 0;
-    let numberOfNormalQuestions = 0;
-    let numberOfHardQuestions = 0;
-    let forPromise = new Promise(async (resolve1, reject1) => {
-      for (let question of bank.idQuestions) {
-        let promise = new Promise((resolve, reject) => {
-          Question.findOne({ _id: question }).exec((err, question) => {
-            if (question) {
-              question.level === "easy"
-                ? resolve("easy")
-                : question.level === "normal"
-                ? resolve("normal")
-                : resolve("hard");
-            }
-          });
-        });
-        await promise
-          .then((res) => {
-            res == "easy"
-              ? numberOfEasyQuestions++
-              : res == "normal"
-              ? numberOfNormalQuestions++
-              : numberOfHardQuestions++;
-          })
-          .catch((err) => {
-            console.log(err);
-          });
-      }
-      resolve1({
-        numberOfEasyQuestions: numberOfEasyQuestions,
-        numberOfNormalQuestions: numberOfNormalQuestions,
-        numberOfHardQuestions: numberOfHardQuestions,
-      });
-    });
-    let obj = {};
-
-    await forPromise.then((res) => {
-      obj = res;
-    });
-
-    bankRes = {
-      ...bank._doc,
-      numberOfEasyQuestions: obj.numberOfEasyQuestions,
-      numberOfNormalQuestions: obj.numberOfNormalQuestions,
-      numberOfHardQuestions: obj.numberOfHardQuestions,
-    };
-
-    // console.log(bankRes);
+  Bank.findOne({ _id: id, idOwner: userId }).exec((err, bank) => {
     if (bank) {
-      res.status(200).send({
-        success: true,
-        message: "[SUCCESS] Get bank info",
-        data: bankRes,
+      console.log(bank);
+      Question.aggregate([
+        {
+          $match: {
+            idBank: id,
+          },
+        },
+        {
+          $project: {
+            level: 1,
+          },
+        },
+        {
+          $group: {
+            _id: "$level",
+            qty: { $sum: 1 },
+          },
+        },
+      ]).exec((err, questions) => {
+        let bankRes = {
+          ...bank._doc,
+        };
+
+        for (q of questions) {
+          q._id === "hard"
+            ? (bankRes.numberOfHardQuestions = q.qty)
+            : q._id === "normal"
+            ? (bankRes.numberOfNormalQuestions = q.qty)
+            : (bankRes.numberOfEasyQuestions = q.qty);
+        }
+
+        res.status(200).send({
+          success: true,
+          message: "[SUCCESS] Get bank info",
+          data: bankRes,
+        });
       });
     } else {
       res.status(202).send({
@@ -105,16 +102,79 @@ module.exports.getOneBank = (req, res) => {
       });
     }
   });
+  // Bank.findOne({ _id: id, idOwner: userId }).exec(async (err, bank) => {
+  //   let bankRes = {};
+  //   let numberOfEasyQuestions = 0;
+  //   let numberOfNormalQuestions = 0;
+  //   let numberOfHardQuestions = 0;
+  //   let forPromise = new Promise(async (resolve1, reject1) => {
+  //     for (let question of bank.idQuestions) {
+  //       let promise = new Promise((resolve, reject) => {
+  //         Question.findOne({ _id: question }).exec((err, question) => {
+  //           if (question) {
+  //             question.level === "easy"
+  //               ? resolve("easy")
+  //               : question.level === "normal"
+  //               ? resolve("normal")
+  //               : resolve("hard");
+  //           }
+  //         });
+  //       });
+  //       await promise
+  //         .then((res) => {
+  //           res == "easy"
+  //             ? numberOfEasyQuestions++
+  //             : res == "normal"
+  //             ? numberOfNormalQuestions++
+  //             : numberOfHardQuestions++;
+  //         })
+  //         .catch((err) => {
+  //           console.log(err);
+  //         });
+  //     }
+  //     resolve1({
+  //       numberOfEasyQuestions: numberOfEasyQuestions,
+  //       numberOfNormalQuestions: numberOfNormalQuestions,
+  //       numberOfHardQuestions: numberOfHardQuestions,
+  //     });
+  //   });
+  //   let obj = {};
+
+  //   await forPromise.then((res) => {
+  //     obj = res;
+  //   });
+
+  //   bankRes = {
+  //     ...bank._doc,
+  //     numberOfEasyQuestions: obj.numberOfEasyQuestions,
+  //     numberOfNormalQuestions: obj.numberOfNormalQuestions,
+  //     numberOfHardQuestions: obj.numberOfHardQuestions,
+  //   };
+
+  //   if (bank) {
+  //     res.status(200).send({
+  //       success: true,
+  //       message: "[SUCCESS] Get bank info",
+  //       data: bankRes,
+  //     });
+  //   } else {
+  //     res.status(202).send({
+  //       success: false,
+  //       message: "[ERROR] Id is not existed",
+  //     });
+  //   }
+  // });
 };
 module.exports.postBank = (req, res) => {
   // console.log(req.userId);
   let idQuestions = [];
 
   const { questions } = req.body;
+  const { userId } = req;
 
   let newBank = new Bank({
     title: req.body.title,
-    idOwner: req.userId,
+    idOwner: userId,
   });
 
   for (let question of questions) {
@@ -124,14 +184,15 @@ module.exports.postBank = (req, res) => {
     //   if (answer.isTrue) count++;
     // }
 
-    let count = question.answers.answers.filter(answer => answer.isTrue).length;
-    
+    const count = question.answers.answers.filter((answer) => answer.isTrue)
+      .length;
+
     console.log(count);
-    let newQuestion = new Question({
+    const newQuestion = new Question({
       title: question.title,
       level: question.level,
       answers: question.answers.answers,
-      isManyAnswers: count >= 2 ? true : false,
+      isManyAnswers: count >= 2,
       idBank: newBank._id,
     });
 
@@ -151,8 +212,9 @@ module.exports.postBank = (req, res) => {
 };
 module.exports.deleteBank = (req, res) => {
   const { id } = req.params;
+  const { userId } = req;
 
-  Bank.findByIdAndDelete({ _id: id }).exec((err, bank) => {
+  Bank.findByIdAndDelete({ _id: id, userId: userId }).exec((err, bank) => {
     if (bank) {
       for (let idQuestion of bank.idQuestions) {
         Question.findByIdAndDelete({ _id: idQuestion }).exec(
@@ -177,10 +239,8 @@ module.exports.searchBank = (req, res) => {
   const { userId } = req;
   const { name } = req.body;
   let { page } = req.body;
-  page = parseInt(page);
+  page = +(page);
 
-  console.log(name);
-  console.log(page);
   Bank.find({ idOwner: userId }).exec((err, banks) => {
     let results = [];
     if (banks) {
@@ -206,6 +266,7 @@ module.exports.searchBank = (req, res) => {
 };
 module.exports.searchQuestion = (req, res) => {
   const { name } = req.params;
+  const { userId } = req;
 
   let results = Question.aggregate([
     {
@@ -218,8 +279,35 @@ module.exports.searchQuestion = (req, res) => {
     {
       $match: {
         res: {
-          $gt: -1
-        }
+          $gt: -1,
+        },
+      },
+    },
+    {
+      $set: {
+        idBank: {
+          $toObjectId: "$idBank",
+        },
+      },
+    },
+    {
+      $lookup: {
+        from: "banks",
+        localField: "idBank",
+        foreignField: "_id",
+        as: "bankInfo",
+      },
+    },
+    {
+      $set: {
+        bankInfo: "$bankInfo.idOwner",
+      },
+    },
+    {
+      $match: {
+        bankInfo: {
+          $eq: userId,
+        },
       },
     },
   ]);
@@ -233,25 +321,63 @@ module.exports.searchQuestion = (req, res) => {
 };
 module.exports.getQuestions = (req, res) => {
   const { id } = req.body;
-  const { limitItems } = req.body;
+  let { limitItems } = req.body;
+
+  const { userId } = req;
+
+  const objectId = mongoose.Types.ObjectId(id);
 
   let { page } = req.body;
 
-  page = parseInt(page);
+  page = +page;
+  limitItems = +limitItems;
 
-  if (id)
-    Question.find({ idBank: id }).exec((err, questions) => {
-      let result = questions.slice((page - 1) * limitItems, page * limitItems);
+  Bank.aggregate([
+    {
+      $match: {
+        _id: objectId,
+        idOwner: userId,
+      },
+    },
+  ]).exec((err, bank) => {
+    if (bank) {
+      const skip = (page - 1) * limitItems;
 
-      if (questions.length > 0) {
-        res.status(200).send({
-          success: true,
-          message: "Get bank's questions by id",
-          data: result,
-        });
-      }
-    });
+      Question.aggregate([
+        {
+          $match: {
+            idBank: id,
+          },
+        },
+        {
+          $skip: skip,
+        },
+        {
+          $limit: limitItems,
+        },
+      ]).exec((err, questions) => {
+        if (questions.length > 0) {
+          res.status(200).send({
+            success: true,
+            message: "Get bank's questions by id",
+            data: questions,
+          });
+        } else
+          res.status(202).send({
+            success: false,
+            message: "Bank's questions is null",
+            data: [],
+          });
+      });
+    } else
+      res.status(202).send({
+        success: false,
+        message: "Invalid id",
+        data: [],
+      });
+  });
 };
+
 module.exports.getQuestionPages = (req, res) => {
   const { id } = req.params;
   const { limitItems } = req.params;
@@ -289,19 +415,13 @@ module.exports.postQuestion = (req, res) => {
 
   if (count > 1) isManyAnswers = true;
 
-  let newQuestion = new Question({
-    title: req.body.title,
-    answers: req.body.answers,
-    level: req.body.level,
-    isManyAnswers: isManyAnswers,
-    idBank: req.body.idBank,
-  });
-
+  let newQuestion = new Question(
+    ({ title, answers, level, isManyAnswers, idBank } = req.body)
+  );
   newQuestion.save();
   console.log(newQuestion);
 
   Bank.findOne({ _id: req.body.idBank }).exec((err, bank) => {
-    // console.log(bank);
     bank.idQuestions = [...bank.idQuestions, newQuestion._id];
     bank.save();
   });
@@ -312,55 +432,68 @@ module.exports.postQuestion = (req, res) => {
     data: newQuestion,
   });
 };
-module.exports.updateQuestion = (req, res) => {
-  let count = 0;
-
-  for (let answer of req.body.answers) {
-    if (answer.isTrue) count++;
-  }
+module.exports.updateQuestion = async (req, res) => {
+  const count = req.body.answers.filter((answer) => answer.isTrue).length;
+  const { _id } = req.body;
+  const { userId } = req;
 
   if (count >= 2) {
     req.body.isManyAnswers = true;
   } else req.body.isManyAnswers = false;
 
-  if (!req.body.title) {
-    Question.findOne({ _id: req.body._id }).exec((err, question) => {
-      Bank.findOne({ _id: question.idBank }).exec((err, bank) => {
-        bank.idQuestions = bank.idQuestions.filter(
-          (bank) => bank._id != req.body._id
-        );
-        console.log(bank.idQuestions);
-        bank.save();
-      });
-    });
-    Question.deleteOne({ _id: req.body._id }).exec((err, question) => {});
+  Question.findOne({ _id: _id }).exec((err, question) => {
+    if (question) {
+      if (!req.body.title) {
+        Bank.findOne({ _id: question.idBank }).exec((err, bank) => {
+          if (bank) {
+            if (bank.idOwner === userId) {
+              bank.idQuestions = bank.idQuestions.filter(
+                (bank) => bank._id != _id
+              );
 
-    res.status(200).send({
-      success: true,
-      message: "[DELETE] Delete question done",
-      data: Object.assign({ title: "" }, req.body),
-    });
-  } else {
-    Question.findOneAndUpdate({ _id: req.body._id }, req.body).exec(
-      (err, question) => {
-        if (question) {
-          res.status(200).send({
-            success: true,
-            message: "[UPDATE] Update question done",
-            data: req.body,
-          });
-        }
+              bank.save();
+
+              Question.deleteOne({ _id: _id }).exec((err, success) => {
+                if (success) {
+                  res.status(200).send({
+                    success: true,
+                    message: "[DELETE] Delete question done",
+                    data: Object.assign({ title: "" }, req.body),
+                  });
+                }
+              });
+            }
+          }
+          return;
+        });
+      } else {
+        Question.findOneAndUpdate({ _id: _id }, req.body).exec(
+          (err, question) => {
+            if (question) {
+              res.status(200).send({
+                success: true,
+                message: "[UPDATE] Update question done",
+                data: req.body,
+              });
+            }
+          }
+        );
       }
-    );
-  }
+    } else {
+      res.status(202).send({
+        success: false,
+        message: "Id is invalid",
+      });
+    }
+  });
 };
 module.exports.updateBank = (req, res) => {
   let { id, title } = req.body;
 
   let objectId = mongoose.Types.ObjectId(id);
 
-  Bank.updateOne({ _id: id }, { title: title }).exec((err, bank) => {
-    if (bank) {
+  Bank.updateOne({ _id: id }, { title: title }).exec((err, success) => {
+    if (success) {
       Bank.aggregate([
         {
           $match: {
@@ -368,7 +501,6 @@ module.exports.updateBank = (req, res) => {
           },
         },
       ]).exec((err, bank) => {
-        console.log(bank);
         res.status(200).send({
           success: true,
           message: "Update finish",
